edit all code

+++ app.js
@@ -3,7 +3,7 @@ var express = require("express");
 const bodyParser = require('body-parser');
 var app = express();
 var {Client} = require('pg');
-
+var client;
 
 // pathモジュールをロードし、pathに代入
 const path = require('path');
@@ -14,70 +14,35 @@ var server = app.listen(1234, function(){
     console.log("Node.js is listening to PORT:" + server.address().port);
 });
 
+app.use(express.json());
+app.use(express.urlencoded({ extended: true, limit: '10mb', parameterLimit: 1000000 }));
+
 // 2-2. 1234番ポートに接続したらpublicフォルダの内容を公開する。
 app.use(express.static(path.join(__dirname, 'public')));
 
-//Ajax-1
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true, limit: '10mb', parameterLimit: 1000000 }));
 
+//　属性情報登録処理
 app.post("/savefeature", (req, res) => {
 
-    var client = new Client({
-        database: "gisdb",
-        user: "postgres",
-        password: "53320706",
-        host: "localhost",
-        port: 5433,
-    });
-
-    var str = req.body;
-
-    let tableName = str.tableName;
-    let gid = str.gid;
-    
-    var query = "";
-    var queryAll;
-    
-    for (const [key, value] of Object.entries(str)) {
-        if(key != "tableName" && key != "gid"){
-            if(value != "" && value != null && typeof(value) == "string"){
-                    
-                if(query == ""){
-                    query = "SET " + key + "=" + "'" + value + "'"; 
-                } else {
-                    query = query + ", " + key + "=" + "'" + value + "'"; 
-                } 
-            }
-        }
-    }
-    
-    if(query == ""){
-        return;
-    }
-    
-    queryAll = "UPDATE " + "\"" + tableName + "\"" + query + "WHERE gid = " + gid;
-    
-    console.log(queryAll);
-    client.connect((err) => {
-        if (err) {
-            console.log('error connecting: ' + err.stack);
-            return;
-        } else {
-            client.query(queryAll, function (err, result) {
-            
-                client.end();
-                console.log("Update success");
-                res.send(str);
-            
-            });
-        }
-    });
+    var ope = "update"
+    connectServer(ope, req, res);
 });
 
+
+//　図形削除処理
 app.post("/delete", (req, res) => {
 
-    var client = new Client({
+    var ope = "delete"
+    connectServer(ope, req, res);
+
+});
+
+
+// サーバー通信メソッド
+function connectServer(ope, req, res){
+
+    // サーバー
+    client = new Client({
         database: "gisdb",
         user: "postgres",
         password: "53320706",
@@ -85,17 +50,58 @@ app.post("/delete", (req, res) => {
         port: 5433,
     });
 
+    // 受信データ
     var str = req.body;
+    
+    //　接続テーブル名取得
     let tableName = str.tableName;
+
+    //　データユニークID
     let gid = str.gid;
 
-    console.log(tableName);
-    console.log(gid);
+    // SQLクエリ
+    var query;
 
-    query = "DELETE FROM " + "\"" + tableName + "\"" + "WHERE gid = " + gid;
+    if(ope == "update"){
 
-    console.log("delete");
+        //結合済みクエリ
+        var queryAll;
 
+        // クエリ作成　結合
+        for (const [key, value] of Object.entries(str)) {
+           
+            resitCharNumLimit(key, value);
+    
+            if(key != "tableName" && key != "gid"){
+                if(value != "" && value != null && typeof(value) == "string"){
+                        
+                    if(query == null){
+                        query = "SET " + key + "=" + "'" + value + "'"; 
+                    } else {
+                        query = query + ", " + key + "=" + "'" + value + "'"; 
+                    } 
+                }
+            }
+        }
+    
+        if(query == ""){
+            return;
+        }
+        
+        queryAll = "UPDATE " + "\"" + tableName + "\"" + query + "WHERE gid = " + gid;
+
+        connect(str, res, queryAll);
+
+    } else if(ope == "delete"){
+
+        query = "DELETE FROM " + "\"" + tableName + "\"" + "WHERE gid = " + gid;
+        connect(str, res, query);
+
+    }
+}
+
+//　サーバー通信実施処理
+function connect(str, res, query){
     console.log(query);
     client.connect((err) => {
         if (err) {
@@ -103,15 +109,39 @@ app.post("/delete", (req, res) => {
             return;
         } else {
             client.query(query, function (err, result) {
-            
                 client.end();
-                console.log("Delete success");
                 res.send(str);
-            
             });
         }
     });
+    
+}
 
 
+// 登録文字数　バリデーション
+function resitCharNumLimit(key, value){
 
-});
+    if(key == "n03_001" && value.length > 8){
+    
+        throw(e);
+    } else if(key == "n03_002" && value.length > 10) {
+        throw(e);
+
+    } else if(key == "n03_003" && value.length > 10) {
+        throw(e);
+
+    } else if(key == "n03_004" && value.length > 14) {
+        throw(e);
+
+    } else if(key == "n03_005" && value.length > 10) {
+        throw(e);
+
+    } else if(key == "n03_006" && value.length > 10) {
+        throw(e);
+
+    } else if(key == "n03_007" && value.length > 5) {
+        throw(e);
+
+    }
+
+}
